requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
openai~=0.28.1
streamlit~=1.36.0